[REFACTOR] AuthProvider useAuth 병합
