+++ src/actions/auth.action.ts
@@ -1,4 +1,5 @@
 "use server";
+
 import axiosInstance from "@/libs/axiosInstance";
 import { MemberInfo } from "@/types/member";
 
@@ -7,10 +8,16 @@ export const getAuthInfo = async (): Promise<{
   userInfo?: MemberInfo | null;
 }> => {
   try {
-    const response = await axiosInstance.post("/api/logout");
+    const response = await axiosInstance.get("/api/members/me");
     const user: MemberInfo = response.data;
-    return { loggedIn: true, userInfo: user };
+    if (user && user.memberId) {
+      return { loggedIn: true, userInfo: user };
+    } else {
+      return { loggedIn: false, userInfo: null };
+    }
   } catch (error) {
     return { loggedIn: false, userInfo: null };
   }
 };
+
+export const logout = async () => {};
