+++ src/providers/AuthProvider.tsx
@@ -1,7 +1,16 @@
 "use client";
-import { useAuth } from "@/hooks/useAuth";
+
+import { useRouter } from "next/navigation";
+import React, {
+  createContext,
+  useEffect,
+  useState,
+  ReactNode,
+  useContext,
+} from "react";
+import axios from "axios";
 import { MemberInfo } from "@/types/member";
-import React, { createContext, ReactNode, useContext } from "react";
+import { getAuthInfo } from "@/actions/auth.action";
 
 interface AuthContextProps {
   isLoggedIn: boolean;
@@ -17,14 +26,37 @@ interface AuthProviderProps {
 const AuthContext = createContext<AuthContextProps | undefined>(undefined);
 
 export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
-  const auth = useAuth();
-  return <AuthContext.Provider value={auth}>{children}</AuthContext.Provider>;
+  const [user, setUser] = useState<MemberInfo | null>(null);
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  const router = useRouter();
+
+  useEffect(() => {
+    const checkAuth = async () => {
+      const { loggedIn, userInfo } = await getAuthInfo();
+      setIsLoggedIn(loggedIn);
+      setUser(userInfo!);
+    };
+    checkAuth();
+  }, []);
+
+  const logout = async () => {
+    await axios.post("/api/logout");
+    setIsLoggedIn(false);
+    setUser(null);
+    router.push("/");
+  };
+
+  return (
+    <AuthContext.Provider value={{ isLoggedIn, user, setIsLoggedIn, logout }}>
+      {children}
+    </AuthContext.Provider>
+  );
 };
 
-export const useAuthContext = (): AuthContextProps => {
+export const useAuth = (): AuthContextProps => {
   const context = useContext(AuthContext);
   if (!context) {
-    throw new Error("useAuthContext must be used within an AuthProvider");
+    throw new Error("useAuth must be used within an AuthProvider");
   }
   return context;
 };
